+++ PastelEx/PastelSettings.cs
@@ -9,7 +9,7 @@ public sealed class PastelSettings
 
     /// <summary>
     /// If set to <see langword="true"/>, changes to color or style using the <see cref="PastelEx.Foreground"/>,
-    /// <see cref="PastelEx.Background"/>, and <see cref="PastelEx.Decorations"/> properties, or the <see cref="PastelEx.Reset"/>
+    /// <see cref="PastelEx.Background"/>, and <see cref="PastelEx.Decorations"/> properties, or the <see cref="PastelEx.ResetPalette"/>
     /// methods, will be immediately applied to the console output. However, overall performance might degrade when frequently
     /// modifying any of the previously specified properties or methods. If set to <see langword="false"/>, the updates will only
     /// take effect when using any of the PastelEx methods (not properties) to modify the output.